+++ packages/trading/src/components/mobile/bottomNavBar/bottomNavBar.script.tsx
@@ -1,19 +1,20 @@
-import { modal } from "@orderly.network/ui";
-import { AccountSheetWidget } from "../accountSheet";
-import { useTradingPageContext } from "../../../provider/context";
-import { PortfolioSheetWidget } from "../portfolioSheet";
-import { useTradingLocalStorage } from "../../../provider/useTradingLocalStorage";
 import { useAccount } from "@orderly.network/hooks";
-import { isTestnet } from "@orderly.network/utils";
-import { useMemo } from "react";
 import { useAppContext } from "@orderly.network/react-app";
 
 export const useBottomNavBarScript = () => {
-  
-
   const { wrongNetwork } = useAppContext();
+  const { account, state } = useAccount();
+
+  /** link device, acally wallet not connect */
+  const onDisconnect = async () => {
+    localStorage.removeItem("orderly_selected_chainId");
+    await account.disconnect();
+  };
+
   return {
     wrongNetwork,
+    status: state.status,
+    onDisconnect,
   };
 };