+++ packages/ui-scaffold/src/components/accountMenu/useWidgetBuilder.script.ts
@@ -5,8 +5,7 @@ import {
 } from "@orderly.network/hooks";
 import { WalletConnectorModalId } from "@orderly.network/ui-connector";
 import { modal, toast } from "@orderly.network/ui";
-import { useCallback, useEffect, useMemo } from "react";
-import { useScaffoldContext } from "../scaffoldContext";
+import { useCallback } from "react";
 import { AccountStatusEnum } from "@orderly.network/types";
 import { useAppContext } from "@orderly.network/react-app";
 import { ChainSelectorId } from "@orderly.network/ui-chain-selector";
@@ -14,7 +13,6 @@ import { ChainSelectorId } from "@orderly.network/ui-chain-selector";
 export const useAccountMenu = (): any => {
   const { disconnect, connectedChain } = useWalletConnector();
   const { account, state } = useAccount();
-  const { checkChainSupport } = useScaffoldContext();
   const { connectWallet } = useAppContext();
 
   const [_, { findByChainId }] = useChains();
@@ -106,6 +104,14 @@ export const useAccountMenu = (): any => {
     }
   }, [state, connectedChain]);
 
+  const onDisconnect = async () => {
+    localStorage.removeItem("orderly_selected_chainId");
+    await disconnect({
+      label: state.connectWallet?.name,
+    });
+    await account.disconnect();
+  };
+
   return {
     address: state.address,
     accountState: state,
@@ -113,12 +119,7 @@ export const useAccountMenu = (): any => {
     onCrateAccount,
     onCreateOrderlyKey,
     onOpenExplorer,
-    onDisconnect: async () => {
-      await disconnect({
-        label: state.connectWallet?.name,
-      });
-      await account.disconnect();
-    },
+    onDisconnect,
   } as const;
 };