+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @orderly.network/cli
 
+## 0.3.19
+
+### Patch Changes
+
+- a3becba1a: update
+- Updated dependencies [a3becba1a]
+  - @orderly.network/core@2.0.2
+
+## 0.3.19-alpha.0
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/core@2.0.2-alpha.0
+
 ## 0.3.18
 
 ### Patch Changes