+++ packages/ui-transfer/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @orderly.network/ui-transfer
 
+## 2.0.2
+
+### Patch Changes
+
+- a3becba1a: update
+- Updated dependencies [a3becba1a]
+  - @orderly.network/ui-chain-selector@2.0.2
+  - @orderly.network/ui-connector@2.0.2
+  - @orderly.network/hooks@2.0.2
+  - @orderly.network/types@2.0.2
+  - @orderly.network/react-app@2.0.2
+  - @orderly.network/ui@2.0.2
+  - @orderly.network/utils@2.0.2
+
+## 2.0.2-alpha.0
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/ui-chain-selector@2.0.2-alpha.0
+  - @orderly.network/ui-connector@2.0.2-alpha.0
+  - @orderly.network/hooks@2.0.2-alpha.0
+  - @orderly.network/types@2.0.2-alpha.0
+  - @orderly.network/react-app@2.0.2-alpha.0
+  - @orderly.network/ui@2.0.2-alpha.0
+  - @orderly.network/utils@2.0.2-alpha.0
+
 ## 2.0.1
 
 ### Patch Changes